Single Player Timed Mode + VFX + TweenUtilities
Single Player now has a timed mode, they can attempt to solve as many 24 game problems in a period of time.

Each solve causes a new "orbiting" bit to appear around the display. The orbiting bits detach from the display and land on the ground in front after completion of a round. They despawn after a period of time.

If the player walks near a landed orbiting bit, the solution to attain the bit is revealed.